clean up code, created architecture, pull request fixes
