+++ lib/features/auth/helpers/patterns.dart
@@ -1,0 +1,2 @@
+final emailPattern = RegExp(
+    r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+");
