+++ lib/features/auth/models/user.model.dart
