+++ lib/features/auth/screens/log_in_by_email_screen.dart
@@ -1,8 +1,9 @@
-import 'package:dog_friends/user_profile_screen.dart';
 import 'package:flutter/material.dart';
 
-class LogInByPhoneScreen extends StatelessWidget {
-  const LogInByPhoneScreen({super.key});
+import '../../user/screens/user_profile_screen.dart';
+
+class LogInByEmailScreen extends StatelessWidget {
+  const LogInByEmailScreen({super.key});
 
   void goToUserProfile(BuildContext context) {
     Navigator.of(context).push(
