+++ lib/features/auth/screens/welcome_screen.dart
@@ -1,5 +1,6 @@
-import 'package:dog_friends/screens/auth_screen.dart';
 import 'package:flutter/material.dart';
+
+import 'auth_screen.dart';
 
 class WelcomeScreen extends StatelessWidget {
   const WelcomeScreen({Key? key}) : super(key: key);
