+++ lib/features/dog/dog_model/dog_model.dart
