+++ lib/features/dog/helper/text_scaler.dart
